requirements.txt
scrapy==2.11.2
psycopg2-binary==2.9.10
streamlit==1.41.1
pandas==2.2.2
numpy==2.1.3
altair==5.5.0
plotly.express==0.4.1
seaborn==0.13.2
matplotlib==3.9.2